pyproject.toml
[tool.poetry]
name = "pdf-csv"
version = "0.1.0"
description = ""
authors = ["saulo-duarte <saulolduarte@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
camelot-py = "^0.11.0"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
ghostscript = "^0.7"
pandas = "^2.2.2"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.32"
unidecode = "^1.3.8"
pypdf2 = "= 2.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
